+++ libft/src/mem/ft_bzero.c
@@ -14,5 +14,5 @@
 
 void ft_bzero(void *s, size_t c) {
     if (c > 0)
-        ft_memset(s, 0, (int)c);
+        ft_memset(s, 0, c);
 }
